Fixed a few window functions for mac, added mouse move support, catch the clown now works on mac
git-svn-id: https://enigma-dev.svn.sourceforge.net/svnroot/enigma-dev/trunk@418 726d713c-bd1a-46cb-9ef0-1b6e075168bb

+++ ENIGMAsystem/SHELL/Platforms/Cocoa/CocoaWindow.h
@@ -25,6 +25,7 @@
  **                                                                              **
  \********************************************************************************/
 #include <string>
+
 using std::string;
 
 void gmw_init();
@@ -69,7 +70,7 @@ int window_get_height();
 
 //Setters
 int window_set_position(double x,double y);
-int window_set_size(double w,double h);
+int window_set_size(unsigned int w,unsigned int h);
 int window_set_rectangle(double x,double y,double w,double h);
 
 //Center
